refactor(providers): move adapters into provider leaves
