+++ self/subcortex/providers/src/provider-adapters.ts
@@ -1,7 +1,7 @@
 export * from './generated/provider-adapters.generated.js';
 
-export { createAnthropicAdapter } from './adapters/anthropic-adapter.js';
+export { createAnthropicAdapter } from './providers/anthropic/adapter.js';
 export {
   createOllamaAdapter,
   isToolCapableModel,
-} from './adapters/ollama-adapter.js';
+} from './providers/ollama/adapter.js';
