+++ self/subcortex/providers/src/providers/anthropic/adapter.ts
@@ -1,4 +1,514 @@
-export {
-  anthropicAdapter as providerAdapter,
-  createAnthropicAdapter,
-} from '../../adapters/anthropic-adapter.js';
+/**
+ * Anthropic provider adapter — native tool-use, cache boundaries, extended thinking.
+ *
+ * ProviderAdapter for the Anthropic Messages API.
+ */
+import type { ILogChannel, TraceId } from '@nous/shared';
+import type { ParsedModelOutput } from '../../shared/output.js';
+import {
+  defineProviderAdapter,
+  type AdapterCapabilities,
+  type AdapterFormatInput,
+  type AdapterFormattedRequest,
+  type ProviderAdapter,
+} from '../../shared/adapter-types.js';
+
+const ANTHROPIC_CAPABILITIES: AdapterCapabilities = {
+  nativeToolUse: true,
+  cacheControl: true,
+  extendedThinking: true,
+  streaming: true,
+};
+
+// ── Content type detection (mirrors output-parser.ts logic) ─────────
+
+const OPENUI_PREFIX = '%%openui\n';
+
+const CARD_TAG_PATTERNS = [
+  '<StatusCard',
+  '<ActionCard',
+  '<ApprovalCard',
+  '<WorkflowCard',
+  '<FollowUpBlock',
+];
+
+function detectContentType(response: string): {
+  response: string;
+  contentType: 'text' | 'openui';
+} {
+  let stripped = response;
+  let hadPrefix = false;
+  if (response.startsWith(OPENUI_PREFIX)) {
+    stripped = response.slice(OPENUI_PREFIX.length);
+    hadPrefix = true;
+  }
+  const hasCardTag = CARD_TAG_PATTERNS.some((p) => stripped.includes(p));
+  if (hadPrefix || hasCardTag) {
+    return { response: stripped, contentType: 'openui' };
+  }
+  return { response, contentType: 'text' };
+}
+
+// ── Format helpers ──────────────────────────────────────────────────
+
+interface AnthropicSystemSegment {
+  type: 'text';
+  text: string;
+  cache_control?: { type: 'ephemeral' };
+}
+
+const TOP_LEVEL_COMBINATOR_KEYS = ['oneOf', 'anyOf', 'allOf'] as const;
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return value !== null && typeof value === 'object' && !Array.isArray(value);
+}
+
+function asStringArray(value: unknown): string[] {
+  return Array.isArray(value) && value.every((item) => typeof item === 'string')
+    ? value
+    : [];
+}
+
+function uniqueStrings(values: readonly string[]): string[] {
+  return Array.from(new Set(values));
+}
+
+function uniqueUnknown(values: readonly unknown[]): unknown[] {
+  const seen = new Set<string>();
+  const result: unknown[] = [];
+  for (const value of values) {
+    const key = JSON.stringify(value);
+    if (seen.has(key)) continue;
+    seen.add(key);
+    result.push(value);
+  }
+  return result;
+}
+
+function enumValues(schema: Record<string, unknown>): unknown[] {
+  if (Array.isArray(schema.enum)) return schema.enum;
+  if (Object.prototype.hasOwnProperty.call(schema, 'const')) return [schema.const];
+  return [];
+}
+
+function mergePropertySchema(
+  existing: unknown,
+  incoming: unknown,
+): unknown {
+  if (!isRecord(existing)) return incoming;
+  if (!isRecord(incoming)) return existing;
+
+  const existingEnums = enumValues(existing);
+  const incomingEnums = enumValues(incoming);
+  if (existingEnums.length > 0 || incomingEnums.length > 0) {
+    const merged = { ...existing, ...incoming };
+    delete merged.const;
+    merged.enum = uniqueUnknown([...existingEnums, ...incomingEnums]);
+    if (existing.type && !incoming.type) merged.type = existing.type;
+    return merged;
+  }
+
+  return { ...existing, ...incoming };
+}
+
+function commonRequired(requiredSets: readonly string[][]): string[] {
+  if (requiredSets.length === 0) return [];
+  return requiredSets.slice(1).reduce(
+    (common, required) => common.filter((name) => required.includes(name)),
+    [...requiredSets[0]],
+  );
+}
+
+function flattenTopLevelCombinators(
+  schema: Record<string, unknown>,
+): { schema: Record<string, unknown>; flattened: boolean } {
+  const combinatorEntries = TOP_LEVEL_COMBINATOR_KEYS
+    .map((key) => [key, schema[key]] as const)
+    .filter((entry): entry is readonly [typeof TOP_LEVEL_COMBINATOR_KEYS[number], unknown[]] =>
+      Array.isArray(entry[1]),
+    );
+
+  if (combinatorEntries.length === 0) {
+    return { schema, flattened: false };
+  }
+
+  const normalized: Record<string, unknown> = { ...schema };
+  for (const key of TOP_LEVEL_COMBINATOR_KEYS) {
+    delete normalized[key];
+  }
+
+  const properties: Record<string, unknown> = isRecord(normalized.properties)
+    ? { ...normalized.properties }
+    : {};
+  const anyOfRequiredSets: string[][] = [];
+  const allOfRequired: string[] = [];
+
+  for (const [key, branches] of combinatorEntries) {
+    for (const branch of branches) {
+      if (!isRecord(branch)) continue;
+
+      if (isRecord(branch.properties)) {
+        for (const [propertyName, propertySchema] of Object.entries(branch.properties)) {
+          properties[propertyName] = propertyName in properties
+            ? mergePropertySchema(properties[propertyName], propertySchema)
+            : propertySchema;
+        }
+      }
+
+      const required = asStringArray(branch.required);
+      if (key === 'allOf') {
+        allOfRequired.push(...required);
+      } else if (required.length > 0) {
+        anyOfRequiredSets.push(required);
+      }
+    }
+  }
+
+  if (Object.keys(properties).length > 0) {
+    normalized.properties = properties;
+  }
+
+  const existingRequired = asStringArray(normalized.required);
+  const required = uniqueStrings([
+    ...existingRequired,
+    ...allOfRequired,
+    ...commonRequired(anyOfRequiredSets),
+  ]);
+  if (required.length > 0) {
+    normalized.required = required;
+  } else {
+    delete normalized.required;
+  }
+
+  if (!normalized.type) {
+    normalized.type = 'object';
+  }
+
+  return { schema: normalized, flattened: true };
+}
+
+function normalizeAnthropicToolInputSchema(
+  toolName: string,
+  inputSchema: Record<string, unknown>,
+  log?: ILogChannel,
+): Record<string, unknown> {
+  const { schema: flattenedSchema, flattened } = flattenTopLevelCombinators(inputSchema);
+  const schema = flattenedSchema.type
+    ? flattenedSchema
+    : { ...flattenedSchema, type: 'object' };
+
+  if (flattened) {
+    log?.info(
+      `Flattening top-level JSON Schema combinator for Anthropic tool "${toolName}"`,
+    );
+  }
+
+  if (!flattenedSchema.type) {
+    log?.info(
+      `Injecting type:"object" for tool "${toolName}" — inputSchema missing type field`,
+    );
+  }
+
+  return schema;
+}
+
+function formatSystemPrompt(
+  systemPrompt: string | string[],
+): string | AnthropicSystemSegment[] {
+  if (typeof systemPrompt === 'string') {
+    return systemPrompt;
+  }
+
+  // String array — cache boundary composition
+  return systemPrompt.map((segment, index) => {
+    const seg: AnthropicSystemSegment = { type: 'text', text: segment };
+    // Cache control on the last segment (longest cache prefix — Anthropic convention)
+    if (index === systemPrompt.length - 1) {
+      seg.cache_control = { type: 'ephemeral' };
+    }
+    return seg;
+  });
+}
+
+function formatTools(
+  toolDefinitions?: readonly import('@nous/shared').ToolDefinition[],
+  log?: ILogChannel,
+): Array<{ name: string; description: string; input_schema: Record<string, unknown> }> | undefined {
+  if (!toolDefinitions || toolDefinitions.length === 0) return undefined;
+
+  return toolDefinitions.map((tool) => {
+    const schema = normalizeAnthropicToolInputSchema(
+      tool.name,
+      (tool.inputSchema as Record<string, unknown>) ?? {},
+      log,
+    );
+    return {
+      name: tool.name,
+      description: tool.description ?? '',
+      input_schema: schema,
+    };
+  });
+}
+
+type AnthropicMessage = { role: 'user' | 'assistant'; content: string | Array<Record<string, unknown>> };
+
+function formatMessages(
+  context: readonly import('@nous/shared').GatewayContextFrame[],
+  log?: ILogChannel,
+): AnthropicMessage[] {
+  const raw: AnthropicMessage[] = context.map((frame) => {
+    // Tool result with tool_call_id metadata → Anthropic tool_result content block
+    if (frame.role === 'tool' && frame.metadata?.tool_call_id) {
+      return {
+        role: 'user' as const,
+        content: [
+          {
+            type: 'tool_result',
+            tool_use_id: frame.metadata.tool_call_id as string,
+            content: frame.content,
+          },
+        ],
+      };
+    }
+
+    // Assistant frame with tool_calls metadata → Anthropic content blocks with tool_use
+    if (frame.role === 'assistant' && Array.isArray(frame.metadata?.tool_calls)) {
+      const contentBlocks: Array<Record<string, unknown>> = [];
+      if (frame.content.trim()) {
+        contentBlocks.push({ type: 'text', text: frame.content });
+      }
+      for (const tc of frame.metadata!.tool_calls as Array<{ id?: string; name: string; input: unknown }>) {
+        if (tc.id) {
+          contentBlocks.push({
+            type: 'tool_use',
+            id: tc.id,
+            name: tc.name,
+            input: tc.input ?? {},
+          });
+        }
+      }
+      if (contentBlocks.length > 0) {
+        return {
+          role: 'assistant' as const,
+          content: contentBlocks,
+        };
+      }
+    }
+
+    return {
+      role: (frame.role === 'tool' || frame.role === 'system' ? 'user' : frame.role) as 'user' | 'assistant',
+      content: frame.content,
+    };
+  });
+
+  // Merge consecutive same-role messages. Anthropic requires all tool_result
+  // blocks for a given assistant tool_use turn to appear in a single user
+  // message immediately after.
+  const merged: AnthropicMessage[] = [];
+  for (const msg of raw) {
+    const prev = merged[merged.length - 1];
+    if (prev && prev.role === msg.role) {
+      const prevBlocks = Array.isArray(prev.content)
+        ? prev.content
+        : [{ type: 'text', text: prev.content }];
+      const curBlocks = Array.isArray(msg.content)
+        ? msg.content
+        : [{ type: 'text', text: msg.content }];
+      prev.content = [...prevBlocks, ...curBlocks];
+    } else {
+      merged.push({ ...msg });
+    }
+  }
+
+  // Debug: log message structure for tool use debugging
+  log?.debug('formatMessages output', {
+    messageCount: merged.length,
+    messages: merged.map((m, i) => {
+      const blocks = Array.isArray(m.content) ? m.content : [];
+      const detail: Record<string, unknown> = {
+        index: i,
+        role: m.role,
+        contentType: Array.isArray(m.content) ? 'blocks' : 'string',
+        blockTypes: blocks.map((b: Record<string, unknown>) => b.type ?? 'text'),
+      };
+      // Show tool_use ids and tool_result tool_use_ids for pairing diagnosis
+      const toolUseIds = blocks
+        .filter((b: Record<string, unknown>) => b.type === 'tool_use')
+        .map((b: Record<string, unknown>) => b.id);
+      const toolResultIds = blocks
+        .filter((b: Record<string, unknown>) => b.type === 'tool_result')
+        .map((b: Record<string, unknown>) => b.tool_use_id);
+      if (toolUseIds.length > 0) detail.toolUseIds = toolUseIds;
+      if (toolResultIds.length > 0) detail.toolResultIds = toolResultIds;
+      return detail;
+    }),
+  });
+
+  return merged;
+}
+
+// ── Response parsing ────────────────────────────────────────────────
+
+interface AnthropicContentBlock {
+  type?: string;
+  id?: string;
+  text?: string;
+  name?: string;
+  input?: unknown;
+  thinking?: string;
+}
+
+interface AnthropicResponse {
+  content?: AnthropicContentBlock[];
+  stop_reason?: string;
+  usage?: { input_tokens?: number; output_tokens?: number };
+}
+
+function parseAnthropicResponse(
+  output: unknown,
+): ParsedModelOutput {
+  if (typeof output === 'string') {
+    // Plain string — treat as text response
+    const detected = detectContentType(output);
+    return {
+      response: detected.response,
+      toolCalls: [],
+      memoryCandidates: [],
+      contentType: detected.contentType,
+    };
+  }
+
+  if (!output || typeof output !== 'object') {
+    return {
+      response: String(output ?? ''),
+      toolCalls: [],
+      memoryCandidates: [],
+      contentType: 'text',
+    };
+  }
+
+  const obj = output as AnthropicResponse;
+  const content = obj.content;
+
+  if (!Array.isArray(content)) {
+    // No content blocks — check for direct text
+    if ('text' in (output as Record<string, unknown>)) {
+      const text = (output as Record<string, unknown>).text;
+      if (typeof text === 'string') {
+        const detected = detectContentType(text);
+        return {
+          response: detected.response,
+          toolCalls: [],
+          memoryCandidates: [],
+          contentType: detected.contentType,
+        };
+      }
+    }
+    return {
+      response: String(output),
+      toolCalls: [],
+      memoryCandidates: [],
+      contentType: 'text',
+    };
+  }
+
+  // Parse content blocks
+  const textParts: string[] = [];
+  const toolCalls: Array<{ name: string; params: unknown; id?: string }> = [];
+  const thinkingParts: string[] = [];
+
+  for (const block of content) {
+    if (!block || typeof block !== 'object') continue;
+
+    if (block.type === 'text' && typeof block.text === 'string') {
+      textParts.push(block.text);
+    } else if (block.type === 'tool_use') {
+      if (typeof block.name === 'string') {
+        toolCalls.push({ name: block.name, params: block.input ?? {}, id: block.id });
+      }
+    } else if (block.type === 'thinking' && typeof block.thinking === 'string') {
+      thinkingParts.push(block.thinking);
+    } else if (block.type === 'thinking' && typeof block.text === 'string') {
+      // Some thinking blocks use `text` field
+      thinkingParts.push(block.text);
+    }
+  }
+
+  // Defensive: stop_reason === 'tool_use' but no tool_use blocks → treat as text-only
+  const response = textParts.join('');
+  const detected = detectContentType(response);
+
+  const result: ParsedModelOutput = {
+    response: detected.response,
+    toolCalls,
+    memoryCandidates: [],
+    contentType: detected.contentType,
+  };
+
+  if (thinkingParts.length > 0) {
+    result.thinkingContent = thinkingParts.join('\n');
+  }
+
+  return result;
+}
+
+// ── Adapter factory ─────────────────────────────────────────────────
+
+export function createAnthropicAdapter(log?: ILogChannel): ProviderAdapter {
+  return {
+    capabilities: ANTHROPIC_CAPABILITIES,
+
+    formatRequest(input: AdapterFormatInput): AdapterFormattedRequest {
+      const system = formatSystemPrompt(input.systemPrompt);
+      const messages = formatMessages(input.context, log);
+      const tools = formatTools(input.toolDefinitions, log);
+
+      const result: Record<string, unknown> = {
+        system,
+        messages,
+      };
+
+      if (tools) {
+        result.tools = tools;
+      }
+
+      // Model requirements pass-through
+      if (input.modelRequirements) {
+        result.model_profile = input.modelRequirements.profile;
+      }
+
+      return { input: result };
+    },
+
+    parseResponse(output: unknown, _traceId: TraceId): ParsedModelOutput {
+      try {
+        return parseAnthropicResponse(output);
+      } catch (error) {
+        // Fallback: never throw from parseResponse
+        log?.error('parseResponse error — falling back to String(output)', {
+          error: error instanceof Error ? { message: error.message, stack: error.stack } : error,
+          outputType: typeof output,
+          outputKeys: output && typeof output === 'object' ? Object.keys(output) : [],
+        });
+        return {
+          response: String(output ?? ''),
+          toolCalls: [],
+          memoryCandidates: [],
+          contentType: 'text',
+        };
+      }
+    },
+  };
+}
+
+export const providerAdapter = defineProviderAdapter({
+  adapterKey: 'anthropic',
+  displayName: 'Anthropic',
+  protocol: 'anthropic-messages',
+  capabilities: ANTHROPIC_CAPABILITIES,
+  create(options) {
+    return createAnthropicAdapter(options?.log);
+  },
+});
+
+export { providerAdapter as anthropicAdapter };
